Isolate things for build
